HermeticFetchContent v1.0.3: HERMETIC_DISCOVER_TARGETS_FILE_PATTERN option for non-default exports
- HERMETIC_DISCOVER_TARGETS_FILE_PATTERN parameter
- Removed unnecessary README_DOC file
- Made get-started example based on main instead of particular version

Older cmake project exports CMake Package Config files and targets export
files manually with a custom name, not using the default cmake
install(EXPORT) scheme.

This adds a parameter for FetchContent_MakeHermetic which allow specifying
a pattern to find these non-default file names.

The patch also ensure that the hfc dependency provider has CMP0057 set to
NEW as if the older targets file name are used, cmake is often on an older
version which doesn't have the right features enabled.