HermeticFetchContent v1.0.4 : No more force checkout of FetchContent projects when unclean (#2)
This replaces the force checkout with a WARNING message informing of local
modifications in the source cache but ultimately permitting it for
debugging purposes.

This on one hand fixes a bug with patches being reset on second configure
after a first failed run, but also on another hand allow developers to modify the source cache to attempt to pass and fix
configuration without the need to push changes online and update the hash
in the FetchContent_Declare.

It also fixes the true and false values in the repositories is_clean
detection.

Change-Id: I0706eed78eeaaa972545ef1038d0bc011a31ad0e