Restoring the test ci to working order
For the various tests to work on ci with cmake-re, we need to connect to the tipi service.
To do this, we'll use a new user. This new user can connect to the tipi/cmake-re service. This user has the minimum rights to use this service and pass the tests.

Furthermore, we know that some tests fail due to bugs in cmake-re or HFC, so for the time being we must mark them as having errors so that we can correct them in the future.
Change-Id: I0706eed78ddbbb971545ef1038d0bc022a32ad0e