HermeticFetchContent v1.0.9 : Improved FORCE_SYSTEM_* behavior
Improve the behavior of hfc_make_available_single when FORCE_SYSTEM_* is
set for a given dependency by forcing find_xxx to use the system root for
it's search.

Additionally the system was made more reliable by ensuring that a proxy
toolchain is generated to forward necessary project information in
addition to enabling the same CMAKE_LANGUAGEs as the invoking project.

Furthermore the forwarding of the main project's CMAKE_MODULE_PATH
information to the proxy toolchain used to generate the hermetic target
cache for a FORCE_SYSTEM_<content-name> was implemented to improve
expectations of behavior.

Namely if an user had a `find_package(XXX REQUIRED)` before and this was
accomplished thanks to the project own CMake FindXXX module, the call to
HermeticFetchContent_MakeAvailableAt* when FORCE_SYSTEM_<content-name> is
set will now be fully equivalent to the find_package, also leveraging the
FindXXX module of the project, and not just the CMake default modules.

Co-authored-by: Yannic Staudt <yannic@tipi.build>
Co-authored-by: Damien Buhl <damien@tipi.build>

Change-Id: Ia55263361f376fb40a7c4b66dc559c7bbc931feb