Improved FORCE_SYSTEM_<content-name> behavior
Improve the behavior of hfc_make_available_single when FORCE_SYSTEM_<content-name> is set for a given dependency by forcing find_xxx to use the system root for its search.
Additionally the system was made more reliable by ensuring that a proxy toolchain is generate to forward necessary project information in addition to enabling the same CMAKE_LANGUAGEs as the invoking project.
Furthermore the forwarding of the main project's CMAKE_MODULE_PATH information to the proxy toolchain of the project generating the target cache for a FORCE_SYSTEM_<content-name> was implemented to improve transparency of behavior.

Co-authored-by: Damien Buhl <damien@tipi.build>

Change-Id: Ia55263361f376fb40a7c4b66dc559c7bbc931feb