HermeticFetchContent v1.0.12 : Fix cmake-re cached project retrieval with HERMETIC_PREPATCHED_RESOLVER
Retrieve the right cache entry when HermeticFetchContent is used with an
HERMETIC_PREPATCHED_RESOLVER, it otherwise would wrongfully use the origin
of the source project as cache-key, which would be correct when the project
get patched during the build but not when a prepatched resolver succeeds.

HERMETIC_PREPATCHED_RESOLVER are meant to help clean builds for non-project
maintainers.

Co-authored-by: Damien Buhl <damien@tipi.build>
Co-authored-by: Luc Lambour <luc@tipi.build>

Change-Id: I0706eed78eeaaa861434ef1038d0bc011431ad0d