HermeticFetchContent v1.0.17 :  HERMETIC_FETCHCONTENT_REMOVE_SOURCE_DIR_AFTER_INSTALL = ON
When a HFC project using the HERMETIC_FETCHCONTENT_REMOVE_SOURCE_DIR_AFTER_INSTALL
had to rebuild the dependency, cloning the source again from a git
repository failed with an unrecoverable error. Fixed by deleting
the tracking subbuild folder when deleting the sources to the state
of source download is effectively reset.

CHANGELOG

- Fix HERMETIC_FETCHCONTENT_REMOVE_SOURCE_DIR_AFTER_INSTALL with HFC source cache

Change-Id: If4072fafe8d771a388885466c9239f6da105b240