HermeticFetchContent v1.0.17 : Fix & Deprecate HERMETIC_FETCHCONTENT_REMOVE_
This deprecates the HERMETIC_FETCHCONTENT_REMOVE_* flags in favor of
HFC_V1_REMOVE_SOURCE_DIR_AFTER_INSTALL. Renaming with a versioned prefix
to ensures users of v1.0.11 and earlier migrate automatically to the source
cache which doesn't remove the source and build trees needlessly.

This fixes bugs with HFC_V1_REMOVE_SOURCE_DIR_AFTER_INSTALL ON on
dependency rebuild.

CHANGELOG

- Fix & Deprecate HERMETIC_FETCHCONTENT_REMOVE_

Co-authored-by: Damien Buhl <damien@tipi.build>
Co-authored-by: Yannic Staudt <yannic@tipi.build>

Change-Id: If4072fafe8d771a388885466c9239f6da105b240