HermeticFetchContent v1.0.18 : Integrate lighter source build for goldilock
Goldilock was building with BUILD_TESTING ON by default which is wasteful
and can come in the users way when porting on platforms where no prebuilt
binary exists.

This hence keeps a minimum version of goldilock v1.2.0 and the released
binaries while it ups the source release to v1.2.1 which is just changing
the build system to exclude tests.

See for more details https://github.com/tipi-build/goldilock/pull/11

Change-Id: I3c831291fce6295326059c1491461341c13ef8e5